+++ R/genericFunctions.R
@@ -250,27 +250,34 @@ control_nb_rows <- function(dataSet, nb_rows, function_name = "", variable_name 
 ###################################################################################################
 # power <- function(x){sum(x^2}
 # Ex: true.aggFunction(c(power = power, sqrt = sqrt))
-true.aggFunction <- function(functions, function_name = "true.aggFunction "){
-  for(fun in names(functions)){
+true.aggFunction <- function(functions, function_name = "true.aggFunction"){
+  for(fun in functions){
     ## Check it
     # check type
-    if (!is.function(functions[[fun]])){
-      warning(paste0(function_name, ": ", fun, " is not a function, it wont be used."))
-      functions <- functions[names(functions) != fun]
+	if (! is.character(fun)){
+		stop(paste0(function_name, ": functions should be a list of names (as character) of functions."))
+	}
+    if (!exists(x = fun)) {
+      warning(paste0(function_name, ": ", fun, " doesn't exist, it wont be used."))
+      functions <- functions[functions != fun]
     }
     else{
-      # check aggregation
-      if (length(functions[[fun]](1:3)) != 1){
-        warning(paste0(function_name, ": ", fun, " is not an aggegaration function, it wont be used. An aggregation function is a function that for multiple input return only one, exemple: sum."))
-        functions <- functions[names(functions) != fun]
+      if (!is.function(get(fun))){
+        warning(paste0(function_name, ": ", fun, " is not a function, it wont be used."))
+        functions <- functions[functions != fun]
       }
-    }
+      else{
+        # check aggregation
+        if (length(get(fun)(1:3)) != 1){
+          warning(paste0(function_name, ": ", fun, " is not an aggregation function, it wont be used. An aggregation function is a function that for multiple input return only one, exemple: sum."))
+          functions <- functions[functions != fun]
+        }
+      }
+    } 
   }
   # Wrapp-up
   return(functions)
 }
-
-