+++ tests/testthat/test_prepareSet.R
@@ -4,7 +4,7 @@ verbose <- TRUE
 # -----------
 data("adult")
 adult <- adult[1:500, ]
-adult_agg <-  prepareSet(adult, key = "country", digits = 2, n_unfactor = 10, verbose = verbose, functions = c(mean, max))
+adult_agg <-  prepareSet(adult, key = "country", digits = 2, n_unfactor = 10, verbose = verbose, functions = c("mean", "max"))
 
 test_that("prepareSet:", 
           {