+++ NEWS.md
@@ -4,6 +4,11 @@ V 0.3.5
 - New features:
 	- New features in existing functions:
 		- *findAndTransFormDates* now as an *ambiguities* parameter, IGNORE to work as before, WARN to check for ambiguities and print them, SOLVE to try to solve ambiguities on more lines. (NB: with SOLVE, *n_test* becomes a not needed user parameter)	
+		- *one_hot_encoder* now uses a *build_encoding* functions to be able to build same encoding on train and on test.
+		
+WARNING:
+- *one_hot_encoder* now requires you to run *build_encoding* first.
+		
 V 0.3.4
 ========
 - Improvement of function